Multi-version: refactor all script-related utils to one module
