+++ extensions/amp-analytics/0.1/config.js
@@ -17,7 +17,7 @@
 import {DEFAULT_CONFIG} from './default-config';
 import {Services} from '../../../src/services';
 import {assertHttpsUrl} from '../../../src/url';
-import {calculateScriptBaseUrl} from '../../../src/service/extension-location';
+import {calculateScriptBaseUrl} from '../../../src/service/extension-script';
 import {deepMerge, dict, hasOwn} from '../../../src/utils/object';
 import {dev, user, userAssert} from '../../../src/log';
 import {getChildJsonConfig} from '../../../src/json';
