+++ src/service/custom-element-registry.js
@@ -17,7 +17,7 @@
 import {ElementStub} from '../element-stub';
 import {Services} from '../services';
 import {createCustomElementClass, stubbedElements} from '../custom-element';
-import {extensionScriptsInNode} from '../element-service';
+import {extensionScriptsInNode} from './extension-script';
 import {reportError} from '../error';
 import {pureUserAssert as userAssert} from '../core/assert';
